ApodizationFunctions

Documentation for ApodizationFunctions.

ApodizationFunctions.BartlettApodization

ApodizationFunctions.BartlettHannApodization

ApodizationFunctions.BlackmanApodization

ApodizationFunctions.CosineApodization

ApodizationFunctions.GaussianApodization

ApodizationFunctions.HammingApodization

ApodizationFunctions.HanningApodization

ApodizationFunctions.KaiserApodization

ApodizationFunctions.LanczosApodization

ApodizationFunctions.RectApodization

ApodizationFunctions.TriangApodization

ApodizationFunctions.TukeyApodization

ApodizationFunctions.BartlettApodization — Type

ApodizationFunctions.BartlettApodization(sz; kwargs...)

Create a N-dimensionnal apodization function ApodizationFunctions.BartlettApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see bartlett function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.BartlettApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.BartlettHannApodization — Type

ApodizationFunctions.BartlettHannApodization(sz; kwargs...)

Create a N-dimensionnal apodization function ApodizationFunctions.BartlettHannApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see bartlett_hann function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.BartlettHannApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.BlackmanApodization — Type

ApodizationFunctions.BlackmanApodization(sz; kwargs...)

Create a N-dimensionnal apodization function ApodizationFunctions.BlackmanApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see blackman function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.BlackmanApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.CosineApodization — Type

ApodizationFunctions.CosineApodization(sz; kwargs...)

Create a N-dimensionnal apodization function ApodizationFunctions.CosineApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see cosine function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.CosineApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.GaussianApodization — Type

ApodizationFunctions.GaussianApodization(sz, par; kwargs...)

Create a a N-dimensionnal apodization function ApodizationFunctions.GaussianApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For par and kwargs definitions see gaussian function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.GaussianApodization((4, 4, 4), 0.5);

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.HammingApodization — Type

ApodizationFunctions.HammingApodization(sz; kwargs...)

Create a N-dimensionnal apodization function ApodizationFunctions.HammingApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see hamming function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.HammingApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.HanningApodization — Type

ApodizationFunctions.HanningApodization(sz; kwargs...)

Create a N-dimensionnal apodization function ApodizationFunctions.HanningApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see hanning function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.HanningApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.KaiserApodization — Type

ApodizationFunctions.KaiserApodization(sz, par; kwargs...)

Create a a N-dimensionnal apodization function ApodizationFunctions.KaiserApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For par and kwargs definitions see kaiser function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.KaiserApodization((4, 4, 4), 0.5);

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.LanczosApodization — Type

ApodizationFunctions.LanczosApodization(sz; kwargs...)

Create a N-dimensionnal apodization function ApodizationFunctions.LanczosApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see lanczos function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.LanczosApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.RectApodization — Type

ApodizationFunctions.RectApodization(sz; kwargs...)

Create a N-dimensionnal apodization function ApodizationFunctions.RectApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see rect function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.RectApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.TriangApodization — Type

ApodizationFunctions.TriangApodization(sz; kwargs...)

Create a N-dimensionnal apodization function ApodizationFunctions.TriangApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For kwargs definitions see triang function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.TriangApodization((4, 4, 4));

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...

ApodizationFunctions.TukeyApodization — Type

ApodizationFunctions.TukeyApodization(sz, par; kwargs...)

Create a a N-dimensionnal apodization function ApodizationFunctions.TukeyApodization, with size sz defined as a single Int (1-dimensionnal) or a tuple of Int.

For par and kwargs definitions see tukey function in package DSP.jl.

Example

julia> apod = ApodizationFunctions.TukeyApodization((4, 4, 4), 0.5);

julia> A = rand(4, 4, 4);

julia> apod(A)
4×4×4 Array{Float64, 3}:
...